ozone/starboard: Enforce "egl" for Ozone Starboard platform (#4607)
This CL adds the command line option --use-gl=egl if not present, or enforces that this is the value passed to the program. This is to avoid falling back to Angle, which is undesired at the moment.

This CL replaces #4601 which had a different approach.

Bug: b/371272304
original-hexsha: a2deb5c325ecea9986913a3afe6c717364f36dd9