[CONFLICT] Original Commit: 8d93a1ddf8500ea2e144b46ddad92157b1d2992a
Original Message:
cobalt/build: Set enable_screen_capture = false to all builds (#5655)

This CL disables building the screen capture stack for all Cobalt builds.
Note that this is different from screenshots, which are dealt with in the
UI compositor. This screen capture is for live streaming purposes, e.g.
for WebRTC.

content/common/features.gni has a bug in the declaration of the variable,
which is also present upstream (https://crrev.com/c/6508106 is up for
review to fix it).

A linux x64x11 devel build is indeed smaller with the flag unset, reduced
from 1,358,454,800B to 1,357,650,928B, i.e. 803,872B, or 803KB, or 0.06%
size reduction. We'll take it : )

Bug: 383140932
original-hexsha: 48abcdc431933758ddf17dfe84c27f80cafd882a


Files with conflicts:
- content/common/features.gni