[POSIX] Add pathconf in Starboard API (#6115)
Add the POSIX symbol pathconf to Starboard. Pathconf is added to
Starboard with a Starboard wrapper that does input translation for the
`name` argument in pathconf.

Testing for pathconf is done through checking whether or not the
specified name configuration is defined and if that configuration has a
limit. If the return value of `pathconf` is -1 but errno is 0, we can
assume that there is no limit for that configuration. If errno is 0 but
the return value is not -1, we then check to see that the specified
value is in accordance to the standard's "minimum acceptable value."
Details on what these values are can be seen here:
https://pubs.opengroup.org/onlinepubs/9699919799/basedefs/limits.h.html

Bug: 416342476
original-hexsha: d9609f1f8cf6d7e9ce07c00851386972eb03ea38