Bringing back network disconnected error dialog popup (#5235)
b/394176365

Brought back the ErrorDialog that popped up when a device does not have
a network connection. Existed in the old Kimono but currently only shows
a blank screen when running causing confusion. Bringing it back helps
the user understand that the issue is with an Internet error.
original-hexsha: 367a73465c66b1b7161b43d97bf49ce15266b8c2