[CONFLICT] Original Commit: 11f52fe6f8a948dc08b3776149e39862457cbacd
Original Message:
BACKPORT: Pass HTMLMediaElement identifier to ::media::Renderer (#5463)

This CL backports
https://chromium-review.googlesource.com/c/chromium/src/+/5528764 to
address the ability for ::media::Renderer to be able to identify
HTMLMediaElements in Chrobalt.

This change is backported from m132+.

The only divergence in this CL are the additions made in node.cc and
node.h. Without these additions, the cherry-pick by itself is not able
to build. These changes include a function declaration and definition
for the `GetDomNodeId` function, along with the include of
`dom_node_id.h`.

Bug: [b/411506193](https://b.corp.google.com/issues/411506193)

Original Commit Description:
By passing an ID to the ::media::Renderer, it allows the Renderer
knowledge about when to apply per playback platform specific
configurations.

One example usage would be to expose configuration to the web app, so it
can control per playback options.

Bug: [b/326658232](https://issuetracker.google.com/326658232)
original-hexsha: 134f771326d6f3f9cfaf74223e909182573dd3b8

Files with conflicts:
- content/renderer/media/media_factory.cc
- content/renderer/media/media_factory.h
- fuchsia_web/webengine/renderer/web_engine_content_renderer_client.cc
- third_party/blink/renderer/core/dom/node.cc
- third_party/blink/renderer/core/dom/node.h
- third_party/blink/renderer/core/exported/web_media_player_impl_unittest.cc
- third_party/blink/renderer/core/html/media/html_media_element.h